remove dependabot
Using an external web service to manage dependencies automatically is not a good idea:

It breaks gradle's reproducible builds feature. Without this,
newcomers are unable to easily check out the project, build and run it.
You essentially drop them back in 90s-style dependency hell,
searching for a combination of dependency versions that work together.

It introduces another external dependency on a corporate product
(dependabot is maintained by GitHub, which is owned by microsoft)
 that may be discontinued at any time

Unattended dependency updates *will* break the project as the dependency's APIs evolve.
This would require developer intervention to update our code to work with the new APIs.
In a FLOSS hobby project like this,
relying on constant developer vigilance to stop the project from
becoming unbuildable for newcomers is not a reasonable expectation.

Far better to let devs choose when to update dependencies themselves,
and carry out the corresponding API change refactoring at the same time.

dependabot's support for Gradle is limited anyway (source:
https://docs.github.com/en/code-security/supply-chain-security/keeping-your-dependencies-updated-automatically/configuration-options-for-dependency-updates#package-ecosystem ),
so using it with Gradle for such a complex build may not be a good idea.